Add instructions for initializing a DB for testing
Developers will need an initialized Trac database in order to run a Trac
instance. The Trac database is usually initialized when the Trac
environment is created with `trac-admin initenv`. There is no official
way to to it later.

In order to get a SQLite DB for testing, run `trac-admin initenv` on
a temporary directory with a DB connection string pointing to a path in
our Trac environment. This is supposed to get interested developers to
a working setup in the easiest way possible.